# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Keuangan Simpadu API",
        "description": "Dokumentasi API Keuangan Simpadu",
        "version": "1.0.0"
    },
    "paths": {
        "/api/kategori-ukt": {
            "get": {
                "tags": [
                    "Kategori UKT"
                ],
                "summary": "Tampilkan semua kategori UKT",
                "operationId": "71f4da472feba4eb685517db057b319c",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id_kategori_ukt": {
                                                "type": "integer"
                                            },
                                            "id_prodi": {
                                                "type": "integer"
                                            },
                                            "kategori_ukt": {
                                                "type": "string"
                                            },
                                            "nominal": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Kategori UKT"
                ],
                "summary": "Tambah kategori UKT baru",
                "operationId": "beaeb8bb41ce33e1002a19bc50497e18",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_prodi",
                                    "kategori_ukt",
                                    "nominal"
                                ],
                                "properties": {
                                    "id_prodi": {
                                        "type": "integer"
                                    },
                                    "kategori_ukt": {
                                        "type": "string"
                                    },
                                    "nominal": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kategori UKT berhasil ditambahkan"
                    }
                }
            }
        },
        "/api/kategori-ukt/{id}": {
            "get": {
                "tags": [
                    "Kategori UKT"
                ],
                "summary": "Detail kategori UKT",
                "operationId": "20eacc948b79445a2cc5aadcac641f3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_kategori_ukt": {
                                            "type": "integer"
                                        },
                                        "id_prodi": {
                                            "type": "integer"
                                        },
                                        "nama_prodi": {
                                            "type": "string"
                                        },
                                        "jenjang": {
                                            "type": "string"
                                        },
                                        "kategori_ukt": {
                                            "type": "string"
                                        },
                                        "nominal": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Kategori UKT"
                ],
                "summary": "Update kategori UKT",
                "operationId": "22ec17b4c8a124f3a6a4a7fedfc25749",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_prodi": {
                                        "type": "integer"
                                    },
                                    "kategori_ukt": {
                                        "type": "string"
                                    },
                                    "nominal": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kategori UKT berhasil diupdate"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Kategori UKT"
                ],
                "summary": "Hapus kategori UKT",
                "operationId": "a8ec8e51e4ec1167be2e96475a08a691",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kategori UKT berhasil dihapus"
                    }
                }
            }
        },
...
    },
...
}